Separate file for ImmutableDeserializedPacket type
