+++ core/src/unprocessed_packet_batches.rs
@@ -1,80 +1,22 @@
 use {
-    crate::transaction_priority_details::{
-        GetTransactionPriorityDetails, TransactionPriorityDetails,
+    crate::{
+        immutable_deserialized_packet::{DeserializedPacketError, ImmutableDeserializedPacket},
+        transaction_priority_details::TransactionPriorityDetails,
     },
     min_max_heap::MinMaxHeap,
     solana_perf::packet::{Packet, PacketBatch},
     solana_sdk::{
         feature_set,
         hash::Hash,
-        message::Message,
-        sanitize::SanitizeError,
-        short_vec::decode_shortu16_len,
-        signature::Signature,
-        transaction::{
-            AddressLoader, SanitizedTransaction, SanitizedVersionedTransaction, Transaction,
-            VersionedTransaction,
-        },
+        transaction::{AddressLoader, SanitizedTransaction, Transaction},
     },
     std::{
         cmp::Ordering,
         collections::{hash_map::Entry, HashMap},
-        mem::size_of,
         rc::Rc,
         sync::Arc,
     },
-    thiserror::Error,
 };
-
-#[derive(Debug, Error)]
-pub enum DeserializedPacketError {
-    #[error("ShortVec Failed to Deserialize")]
-    // short_vec::decode_shortu16_len() currently returns () on error
-    ShortVecError(()),
-    #[error("Deserialization Error: {0}")]
-    DeserializationError(#[from] bincode::Error),
-    #[error("overflowed on signature size {0}")]
-    SignatureOverflowed(usize),
-    #[error("packet failed sanitization {0}")]
-    SanitizeError(#[from] SanitizeError),
-    #[error("transaction failed prioritization")]
-    PrioritizationFailure,
-}
-
-#[derive(Debug, PartialEq, Eq)]
-pub struct ImmutableDeserializedPacket {
-    original_packet: Packet,
-    transaction: SanitizedVersionedTransaction,
-    message_hash: Hash,
-    is_simple_vote: bool,
-    priority_details: TransactionPriorityDetails,
-}
-
-impl ImmutableDeserializedPacket {
-    pub fn original_packet(&self) -> &Packet {
-        &self.original_packet
-    }
-
-    pub fn transaction(&self) -> &SanitizedVersionedTransaction {
-        &self.transaction
-    }
-
-    pub fn message_hash(&self) -> &Hash {
-        &self.message_hash
-    }
-
-    pub fn is_simple_vote(&self) -> bool {
-        self.is_simple_vote
-    }
-
-    pub fn priority(&self) -> u64 {
-        self.priority_details.priority
-    }
-
-    pub fn compute_unit_limit(&self) -> u64 {
-        self.priority_details.compute_unit_limit
-    }
-}
 
 /// Holds deserialized messages, as well as computed message_hash and other things needed to create
 /// SanitizedTransaction
@@ -101,25 +43,10 @@ impl DeserializedPacket {
         packet: Packet,
         priority_details: Option<TransactionPriorityDetails>,
     ) -> Result<Self, DeserializedPacketError> {
-        let versioned_transaction: VersionedTransaction = packet.deserialize_slice(..)?;
-        let sanitized_transaction = SanitizedVersionedTransaction::try_from(versioned_transaction)?;
-        let message_bytes = packet_message(&packet)?;
-        let message_hash = Message::hash_raw_message(message_bytes);
-        let is_simple_vote = packet.meta.is_simple_vote_tx();
-
-        // drop transaction if prioritization fails.
-        let priority_details = priority_details
-            .or_else(|| sanitized_transaction.get_transaction_priority_details())
-            .ok_or(DeserializedPacketError::PrioritizationFailure)?;
+        let immutable_section = ImmutableDeserializedPacket::new(packet, priority_details)?;
 
         Ok(Self {
-            immutable_section: Rc::new(ImmutableDeserializedPacket {
-                original_packet: packet,
-                transaction: sanitized_transaction,
-                message_hash,
-                is_simple_vote,
-                priority_details,
-            }),
+            immutable_section: Rc::new(immutable_section),
             forwarded: false,
         })
     }
@@ -140,18 +67,6 @@ impl Ord for DeserializedPacket {
         self.immutable_section()
             .priority()
             .cmp(&other.immutable_section().priority())
-    }
-}
-
-impl PartialOrd for ImmutableDeserializedPacket {
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl Ord for ImmutableDeserializedPacket {
-    fn cmp(&self, other: &Self) -> Ordering {
-        self.priority().cmp(&other.priority())
     }
 }
 
@@ -387,19 +302,6 @@ pub fn deserialize_packets<'a>(
     packet_indexes.iter().filter_map(move |packet_index| {
         DeserializedPacket::new(packet_batch[*packet_index].clone()).ok()
     })
-}
-
-/// Read the transaction message from packet data
-pub fn packet_message(packet: &Packet) -> Result<&[u8], DeserializedPacketError> {
-    let (sig_len, sig_size) = packet
-        .data(..)
-        .and_then(|bytes| decode_shortu16_len(bytes).ok())
-        .ok_or(DeserializedPacketError::ShortVecError(()))?;
-    sig_len
-        .checked_mul(size_of::<Signature>())
-        .and_then(|v| v.checked_add(sig_size))
-        .and_then(|msg_start| packet.data(msg_start..))
-        .ok_or(DeserializedPacketError::SignatureOverflowed(sig_size))
 }
 
 pub fn transactions_to_deserialized_packets(
